requirements.txt
kopf
kubernetes
requests
pyyaml